requirements.txt
fastapi
uvicorn
google-adk
a2a-python
fastmcp
google-cloud-aiplatform
google-cloud-storage
anthropic[vertex]
